% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcast.R
\name{mdcast}
\alias{mdcast}
\title{Cast a Molten \link[base]{data.frame} with \emph{Multiple} \code{value.var}}
\usage{
mdcast(
  data,
  formula,
  ...,
  value.var = setdiff(names(data), y = all.vars(formula))
)
}
\arguments{
\item{data}{a molten \link[base]{data.frame},
returned object of \link[reshape2]{melt.data.frame}}

\item{formula}{casting \link[stats]{formula}, see function \link[reshape2]{dcast}}

\item{...}{additional parameters of functions \link[reshape2]{acast} and \link[reshape2]{dcast},
which eventually get passed into function \code{reshape2:::cast}.}

\item{value.var}{\link[base]{character} \link[base]{vector},
names of columns which store the values}
}
\value{
Function \code{\link[=mdcast]{mdcast()}} returns a \link[base]{data.frame}.
}
\description{
Cast a molten \link[base]{data.frame} with \emph{multiple} \code{value.var}.
}
\details{
Function \code{\link[=mdcast]{mdcast()}} is an extension of \link[reshape2]{dcast} in the following aspects,

\itemize{

\item {\link{mdcast} handles \strong{m}ultiple \code{value.var}.
For the \eqn{i}-th value variable in \code{value.var},
the \link[reshape2]{acast} columns are named in the fashion of
\verb{value[i].variable[j]}, \eqn{j=1,\cdots,J}.
This is follows naturally from the way \link[base]{data.frame} handles
multiple \link[base]{matrix} input in \code{...}.
}

\item {For the \eqn{i}-th value variable in \code{value.var},
if one-and-only-one of the \link[reshape2]{acast} columns
contains non-\link[base]{missing} elements,
this column is named as \code{value[i]}, instead of \verb{value[i].variable[j]}.
Specifically, we remove the all-\code{NA} columns
from the \link[reshape2]{acast} columns of the \eqn{i}-th value variable.
If one-and-only-one column remains,
we convert this single-column \link[base]{matrix} into a \link[base]{vector}.
We pass this \link[base]{vector} into \link[base]{data.frame}
with the other \link[reshape2]{acast} columns of the other value variables.
This is a super useful feature in practice,
e.g., some measurement only pertains to one of the multiple visits,
therefore we do not need to specify to which visit it corresponds.
}

}
}
\note{
Function \code{\link[=mdcast]{mdcast()}} uses unexported function \code{reshape2:::cast} illegally.
I have asked Hadley, but he has no plan to export \code{reshape2:::cast}.
}
\examples{
library(reshape2)
head(aqm <- melt(airquality, id = c('Month', 'Day'), na.rm = TRUE, value.name = 'v1'))
aqm$v2 = rnorm(dim(aqm)[1L])
head(aqm)
aqm |> 
 dcast(formula = Month + Day ~ variable, value.var = 'v1') |> 
 head()
aqm |> 
 mdcast(formula = Month + Day ~ variable, value.var = c('v1', 'v2')) |>
 head()

(x <- data.frame(
  subj = rep(1:2, each = 3),
  event = rep(paste0('evt', 1:3), times = 2), 
  date = as.Date(c(14001:14003, 18001:18003)),
  y1 = rnorm(6), 
  y2 = c(rnorm(1), NA, NA, rnorm(1), NA, NA),
  y3 = c(rnorm(2), NA, rnorm(2), NA)))
x |> 
 mdcast(formula = subj ~ event, value.var = c('y1', 'y2')) # very useful !!!
x |> 
 mdcast(formula = subj ~ event) # very useful !!!

}
\keyword{internal}
